+++ modules/usrloc/dlist.c
@@ -153,7 +153,7 @@ static int get_domain_db_ucontacts(udomain_t *d, void *buf, int *len,
 #ifdef ORACLE_USRLOC
 	" %.*s, %.*s, %.*s from %s where %.*s > %lu and mod(contact_id, %u) = %u",
 #else
-	" %.*s, %.*s, %.*s from %s where %.*s > %lu and contact_id %% %u = %u",
+	" %.*s, %.*s, %.*s from %s where %.*s > %lld and contact_id %% %u = %u",
 #endif
 		received_col.len, received_col.s,
 		contact_col.len, contact_col.s,
@@ -163,7 +163,7 @@ static int get_domain_db_ucontacts(udomain_t *d, void *buf, int *len,
 		contactid_col.len, contactid_col.s,
 		d->name->s,
 		expires_col.len, expires_col.s,
-		now,
+		(long long)now,
 		part_max, part_idx);
 
 	LM_DBG("query: %.*s\n", (int)(sizeof query_buf), query_buf);
